requirements.txt
-i https://pypi.org/simple
cassandra-driver==3.25.0
click==8.1.3 ; python_version >= '3.7'
geomet==0.2.1.post1 ; python_version >= '2.7' and python_version != '3.3' and python_version < '4'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tabulate==0.9.0
